refactor: modularize csv formatters
